+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   usefull when its IP address got blocked by anti-scraping protection.
 - Added option `liveView: true` to `Apify.launchPuppeteer()` that will start a live view server proving web page
   with overview of all running Puppeteer instances and their screenshots.
+- `PuppeteerPool` now kills opened Chrome instances in `SIGINT` signal.
 
 0.5.42 / 2018-07-04
 ===================